temporal-backend/requirements.txt
openai-agents>=0.2.4
temporalio==1.8.0
python-dotenv==1.0.0